more partial progress
