+++ rtx_package/src/engine.rs
@@ -14,9 +14,9 @@ mod tex_file_io;
 mod tex_inserts;
 mod tex_job;
 mod tex_kern;
-// mod tex_logic;
-// mod tex_macro;
-// mod tex_marks;
+mod tex_logic;
+mod tex_macro;
+mod tex_marks;
 mod tex_math;
 // mod tex_page;
 // mod tex_paragraph;
@@ -42,7 +42,6 @@ mod tex_appendix_b_to_p349;
 mod tex_assignment;
 mod tex_ch24_primitives;
 mod tex_ch25_primitives;
-mod tex_expandable_primitives;
 pub mod base_functions; // auxiliary functions
 mod tex_math_accents;
 mod tex_math_mode;
