+++ rtx_package/src/engine/tex_ch25_primitives.rs
@@ -25,39 +25,6 @@ LoadDefinitions!({
   //     Box($s, undef, undef, Invocation(T_CS('\mkern'), $length),
   //       width => $length, isSpace => 1); });
 
-  // \lower <dimen> <box>
-  // \raise <dimen> <box>
-  // But <box> apparently must really explicitly be an \hbox, \vbox or \vtop (?)
-  // OR something that expands into one!!
-
-  DefConstructor!("\\lower Dimension MoveableBox",
-  // TODO: SVG
-  // "?&inSVG()(<svg:g transform='#transform' _noautoclose='1'>#2</svg:g>)\
-  // (<ltx:text yoffset='#y'  _noautoclose='1'>#2</ltx:text>)",
-  "<ltx:text yoffset='#y'  _noautoclose='1'>#2</ltx:text>",
-    // sizer => sub { raisedSizer($_[0]->getArg(2), $_[0]->getArg(1)->negate); },
-    after_digest => sub[whatsit] {
-      let y         = Dimension(-whatsit.get_arg(1).unwrap().value_of());
-      let ypx       = y.px_value(None);
-      let transform = if ypx != 0.0 { s!("translate(0,{ypx})") } else { String::new() };
-      whatsit.set_property("y", y);
-      whatsit.set_property("transform", transform);
-    }
-  );
-
-  DefConstructor!("\\raise Dimension MoveableBox",
-  // TODO: SVG
-  // "?&inSVG()(<svg:g transform='#transform' _noautoclose='1'>#2</svg:g>)"
-  //   . "(<ltx:text yoffset='#y'  _noautoclose='1'>#2</ltx:text>)",
-  "<ltx:text yoffset='#y'  _noautoclose='1'>#2</ltx:text>",
-  //sizer       => sub { raisedSizer($_[0]->getArg(2), $_[0]->getArg(1)); },
-  after_digest => sub[whatsit] {
-    let y         = Dimension(whatsit.get_arg(1).unwrap().value_of());
-    let ypx       = y.px_value(None);
-    let transform = if ypx != 0.0 { s!("translate(0,{ypx})") } else { String::new() };
-    whatsit.set_property("y", y);
-    whatsit.set_property("transform", transform);
-  });
 
   // Implement ???
   // DefMacro('\vrule','\relax');
