+++ rtx_package/src/engine/tex_macro.rs
@@ -1,0 +1,235 @@
+//! TeX Macro
+//! 
+//! Core TeX Implementation for LaTeXML
+
+use crate::prelude::*;
+LoadDefinitions!({
+  //%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+  // Macro Family of primitive control sequences
+  //%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+  //
+  //======================================================================
+  // Basics
+  //----------------------------------------------------------------------
+  // \begingroup       c  starts a group that must be ended by \endgroup.
+  // \endgroup         c  ends a group that was begun by \begingroup.
+  // \relax            c  is a control sequence which typesets nothing.
+  // \afterassignment  c  saves a token and inserts it after the next assignment.
+  // \aftergroup       c  saves a token and inserts it after the current group is complete.
+  DefPrimitive!("\\begingroup", { begingroup(); });
+  DefPrimitive!("\\endgroup", { endgroup()?; });
+  // This makes \relax disappear completely after digestion
+  // (which seems most TeX like).
+  DefPrimitive!(T_CS!("\\relax"), None, {});
+  //## However, this keeps a box, so it can appear in UnTeX
+  //## DefPrimitive('\relax',undef);
+  //# But if you do that, you've got to watch out since it usually
+  //## shouldn't be a box; See the isRelax code in handleScripts, below
+
+  // NON-STANDARD: Internal token produced by Gullet in response to \dont_expand;
+  // Acts like \relax, but isn't equal to it.
+  DefPrimitive!(T_CS!("\\special_relax"), None, { });
+
+  // \afterassignment saves ONE token (globally!) to execute after the next assignment
+  DefPrimitive!("\\afterassignment Token", sub[(t)] {
+    assign_value("afterAssignment", t, Some(Scope::Global));
+  });
+  // \aftergroup saves ALL tokens (from repeated calls) to be executed IN ORDER after the next
+  // egroup or }
+  DefPrimitive!("\\aftergroup Token", sub[(t)] { push_value("afterGroup", t) });
+
+  //======================================================================
+  // CSName
+  //----------------------------------------------------------------------
+  // \csname           c  forms a control sequence name from the characters making up a collection of tokens.
+  // \endcsname        c  is used with \csname to make a control sequence name.
+  DefParameterType!(CSName, reader => reader!( _inner, _extra, {
+    let mut cs = escapechar();
+    let endcsname_token = T_CS!("\\endcsname");
+    // keep newlines from having \n inside!
+    while let Some(token) = gullet::read_x_token(Some(true), true)? {
+      if token == endcsname_token {
+        break;
+      }
+      match token.get_catcode() {
+        Catcode::CS => {
+          if lookup_definition(&token)?.is_some() {
+            let message =
+              s!("The control sequence {:?} should not appear between \\csname and \\endcsname",
+                token);
+            Error!("unexpected", token, message);
+          } else {
+            let message = s!("The token {:?} is not defined", token);
+            Error!("undefined", token, message);
+          }
+        },
+        Catcode::SPACE => {  // Keep newlines from having \n!
+          cs.push(' ');
+        },
+        _ => {
+          token.with_str(|s| cs.push_str(s));
+        }
+      };
+    }
+    T_CS!(cs)
+  }));
+  DefMacro!("\\csname CSName", sub[(token)] {
+    if lookup_meaning(&token).is_none() {
+      let relax_meaning = lookup_meaning(&TOKEN_RELAX).unwrap();
+      assign_meaning(&token, relax_meaning, None);
+    }
+    token
+  });
+
+  DefPrimitive!("\\endcsname", {
+    Error!("unexpected" ,"\\endcsname", "Extra \\endcsname");
+    //$stomach->getGullet->showUnexpected);
+  });
+
+  //======================================================================
+  // Definition flags
+  //----------------------------------------------------------------------
+  // \global        c  is an assignment prefix which makes the assignment transcend its group.
+  // \long          c  is a prefix for definitions which require multi-paragraph arguments.
+  // \outer         c  is a prefix for a definition which restricts where the definition may be used.
+  // \globaldefs    pi if positive, all assignments are global; if negative, \global is ignored.
+
+  DefPrimitive!("\\global",{ state::set_prefix("global"); }, is_prefix => true);
+  DefPrimitive!("\\long",  { state::set_prefix("long");   }, is_prefix => true);
+  DefPrimitive!("\\outer", { state::set_prefix("outer");  }, is_prefix => true);
+
+  DefRegister!("\\globaldefs", Number!(0));
+  //======================================================================
+  // Definitions
+  //----------------------------------------------------------------------
+  // \def          c  defines a macro.
+  // \edef         c  is similar to \def, except control sequences in the replacement
+  //                  text are expanded when the definition is made.
+  // \gdef         d  is equivalent to `\global\def'.
+  // \xdef         d  is equivalent to `\global\edef'.
+  DefPrimitive!("\\def SkipSpaces Token UntilBrace DefPlain",
+    sub[(cs,params,body)] {
+      do_def(false,cs,params,body)?;
+    },
+    locked => true);
+  DefPrimitive!("\\gdef SkipSpaces Token UntilBrace DefPlain",
+    sub[(cs,params,body)] {
+      do_def(true,cs,params,body)?;
+    },
+    locked => true);
+  DefPrimitive!("\\edef SkipSpaces Token UntilBrace DefExpanded",
+    sub[(cs,params,body)] {
+      do_def(false,cs,params,body)?;
+    },
+    locked => true);
+  DefPrimitive!("\\xdef SkipSpaces Token UntilBrace DefExpanded",
+    sub[(cs,params,body)] {
+      do_def(true,cs,params,body)?;
+    },
+    locked => true);
+  //======================================================================
+  // Copying definitions
+  //----------------------------------------------------------------------
+  //  \let       c  gives a control sequence a token's current meaning.
+  // \futurelet  c  `<cs> <token1> <token2>' is equivalent to `\let <cs> = <token2> <token1> <token2>'.
+  DefPrimitive!("\\let SkipSpaces Token SkipSpaces SkipMatch:= Skip1Space Token", 
+  sub[(token1, token2)] {
+    Let!(token1, token2);
+  });
+  DefPrimitive!("\\futurelet Token Token Token", sub[(cs, token1, token2)] {
+    // NOT expandable, but puts tokens back
+    gullet::unread(Tokens!(token1,token2));
+    Let!(cs, token2);
+  });
+  //======================================================================
+  // Expansion control
+  //----------------------------------------------------------------------
+  // \expandafter      c  `<token1><token2>' is equivalent to `<token1> expansion of <token2>'.
+  // \noexpand         c  prevents the expansion of the following token.
+  DefMacro!("\\expandafter Token Token", sub[(tok, xtok)] {
+    let mut xtok = xtok;
+    let mut skipped : Vec<Token> = vec![tok];
+    while xtok.defined_as(&TOKEN_EXPANDAFTER) {
+      if let Some(ntok) = gullet::read_token()? {
+        skipped.push(ntok);
+        if let Some(nxtok) = gullet::read_token()? {
+          xtok = nxtok;
+        } else {
+          Error!("expected","expandafter", "\\expandafter wrongly used without 2 arguments.");
+        }
+      } else {
+        Error!("expected", "expandafter", "\\expandafter wrongly used without 2 arguments.");
+      }
+    }
+    if let Some(defn) = lookup_expandable(&xtok, false)? {
+      state::local_current_token(xtok);
+      let invoked = defn.invoke(true)?;
+      if !invoked.is_empty() {
+        skipped.extend(invoked.unlist()); // Expand `xtok` ONCE ONLY!
+      }
+      state::expire_current_token();
+    } else if lookup_meaning(&xtok).is_none() {
+      // Undefined token is an error, as expansion is expected.
+      // BUT The unknown token is NOT consumed, (see TeX B book, item 367)
+      // since probably in a real TeX run it would have been defined.
+      state::generate_error_stub(&xtok)?;
+      skipped.push(xtok);
+    } else {
+      skipped.push(xtok);
+    };
+    Ok(Tokens::new(skipped))
+  });
+  // If next token is expandable, prefix it with the internal marker \dont_expand
+  // That token is never defined, explicitly handled in Gullet & should never escape the Gullet
+  DefMacro!(T_CS!("\\noexpand"), None, {
+    if let Some(token) = gullet::read_token()? {
+      let cc = token.get_catcode();
+      if matches!(cc, Catcode::CS | Catcode::ACTIVE) && state::is_dont_expandable(&token) {
+        vec![T_CS!("\\dont_expand"), token]
+      } else {
+        vec![token]
+      }
+    } else {
+      // Missing token likely the result of "{\noexpand}" for which TeX would be unperturbed
+      Vec::new()
+    }
+  });
+  // NON-STANDARD:
+  DefPrimitive!(T_CS!("\\dont_expand"), None, {
+    Error!("misdefined", "\\dont_expand",
+      "The token \\dont_expand should never reach Stomach!"); });
+
+  //======================================================================
+  // \the
+  //----------------------------------------------------------------------
+  // \the              c  returns character tokens for an internal quantity's or parameter's current value.
+  
+  // \the<internal quantity>
+
+  // \the<internal quantity>
+  DefMacro!("\\the Register", sub[args] {
+    let [rdef] : [_; 1] = args.try_into().unwrap();
+    if let ArgWrap::RegisterDefinition(dbox) = rdef {
+      let (rtoken, inner) = *dbox;
+      // let register_type = defn.borrow().register_type;
+      //     if (!$type) {
+      //       my $cs = ToString($defn->getCS);
+      //       Error('unexpected', "\\the$cs", $gullet,
+      //     "You can't use $cs after \\the"); return (); }
+      let defn = rtoken.to_register()
+        .expect("if a Register parameter provides a token, it must have a Register definition.");
+      let value = defn.value_of(inner)
+        .unwrap_or_else(|| RegisterValue::Tokens(Tokens!()));
+      // In all cases, these should be OTHER, except for space. (!?)
+      match value {
+        RegisterValue::Tokens(ts) => ts.unlist(),
+        RegisterValue::Token(t) => vec![t],
+        rv => Explode!(rv.to_string()),
+      }
+    } else {
+      Error!("expected", "<register>", "a register was expected to be here");
+      Vec::new()
+    }
+  });
+  
+});
